requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.10
click==8.0.3
click-log==0.3.2
cloudinary==1.28.1
cloudinary-cli==1.4.6
dj-database-url==0.5.0
Django==3.2.9
django-admin-interface==0.18.4
django-cloudinary-storage==0.3.0
django-colorfield==0.6.3
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-location-field==2.1.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.3
Jinja2==3.0.3
MarkupSafe==2.0.1
Pillow==8.4.0
psycopg2-binary==2.9.3
Pygments==2.11.2
python-decouple==3.5
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.0.0
urllib3==1.26.8
whitenoise==5.3.0
